# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BidCollectionInput {
  /**
     * @minItems 1
     * @maxItems 2000
     * @items.pattern ^R[0-9A-Z]{12}-[0-9]{3}$
     */
  noticeNumbers: string[];
  /**
     * @minItems 1
     * @maxItems 20
     * @items.minLength 1
     * @items.maxLength 100
     */
  keywords: string[];
}

export interface BidImportInput {
  fileName: string;
  contentBase64: string;
}

export interface BidImportResult {
  noticeNumbers: string[];
  duplicateCount: number;
  invalidValues: string[];
}

export type BidCollectionJobStatus = typeof BidCollectionJobStatus[keyof typeof BidCollectionJobStatus];


export const BidCollectionJobStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  completed_with_errors: 'completed_with_errors',
  failed: 'failed',
} as const;

export interface BidCollectionJob {
  jobId: string;
  status: BidCollectionJobStatus;
  totalCount: number;
  completedCount: number;
}

export type BidCollectionStatusStatus = typeof BidCollectionStatusStatus[keyof typeof BidCollectionStatusStatus];


export const BidCollectionStatusStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  completed_with_errors: 'completed_with_errors',
  failed: 'failed',
} as const;

export type BidNoticeResultStatus = typeof BidNoticeResultStatus[keyof typeof BidNoticeResultStatus];


export const BidNoticeResultStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  partial: 'partial',
  failed: 'failed',
} as const;

export type BidAttachmentResultDownloadStatus = typeof BidAttachmentResultDownloadStatus[keyof typeof BidAttachmentResultDownloadStatus];


export const BidAttachmentResultDownloadStatus = {
  pending: 'pending',
  success: 'success',
  failed: 'failed',
} as const;

export interface BidAttachmentResult {
  fileName: string;
  url: string;
  priority: boolean;
  downloadStatus: BidAttachmentResultDownloadStatus;
  /** @nullable */
  downloadError: string | null;
  attempts: number;
  parsedFileCount: number;
  parseFailureCount: number;
  keywordFound: boolean;
}

export interface BidNoticeResult {
  noticeNumber: string;
  /** @nullable */
  noticeName: string | null;
  requestedOrder: string;
  /** @nullable */
  resolvedOrder: string | null;
  /** @nullable */
  orderMatched: boolean | null;
  lookupAttempts: number;
  status: BidNoticeResultStatus;
  attachmentCount: number;
  downloadedCount: number;
  failedCount: number;
  keywordHitCount: number;
  parseFailureCount: number;
  attachments: BidAttachmentResult[];
  /** @nullable */
  error: string | null;
}

export type BidSearchResultDownloadStatus = typeof BidSearchResultDownloadStatus[keyof typeof BidSearchResultDownloadStatus];


export const BidSearchResultDownloadStatus = {
  success: 'success',
  failed: 'failed',
} as const;

export type BidSearchResultParseStatus = typeof BidSearchResultParseStatus[keyof typeof BidSearchResultParseStatus];


export const BidSearchResultParseStatus = {
  success: 'success',
  failed: 'failed',
  not_applicable: 'not_applicable',
} as const;

export type BidSearchResultResultStatus = typeof BidSearchResultResultStatus[keyof typeof BidSearchResultResultStatus];


export const BidSearchResultResultStatus = {
  keyword_found: 'keyword_found',
  keyword_not_found: 'keyword_not_found',
  no_attachment: 'no_attachment',
  download_failed: 'download_failed',
  parse_failed: 'parse_failed',
} as const;

export type BidSearchResultAwardStatus = typeof BidSearchResultAwardStatus[keyof typeof BidSearchResultAwardStatus];


export const BidSearchResultAwardStatus = {
  confirmed: 'confirmed',
  not_found: 'not_found',
} as const;

export interface BidSearchResult {
  noticeNumber: string;
  /** @nullable */
  noticeName: string | null;
  attachmentFileName: string;
  downloadStatus: BidSearchResultDownloadStatus;
  /** @nullable */
  downloadError: string | null;
  parseStatus: BidSearchResultParseStatus;
  /** @nullable */
  parseError: string | null;
  resultStatus: BidSearchResultResultStatus;
  retryCount: number;
  keywordFound: boolean;
  foundKeywords: string[];
  fileName: string;
  /** @nullable */
  sheet: string | null;
  /** @nullable */
  page: number | null;
  location: string;
  surroundingText: string;
  originalText: string;
  itemSpecification?: string;
  itemName?: string;
  itemQuantity?: string;
  itemUnit?: string;
  itemAmount?: string;
  bidderName?: string;
  awardIdentifier?: string;
  awardDate?: string;
  awardAmount?: string;
  bidderAddress?: string;
  bidderPhone?: string;
  budgetAmount?: string;
  estimatedAmount?: string;
  baseAmount?: string;
  constructionOverview?: string;
  awardStatus?: BidSearchResultAwardStatus;
}

export interface BidCollectionSummary {
  notices: number;
  attachments: number;
  downloaded: number;
  downloadFailures: number;
  parseFailures: number;
  lookupFailures: number;
  filesWithKeywords: number;
  keywordMatches: number;
}

export interface BidCollectionStatus {
  jobId: string;
  status: BidCollectionStatusStatus;
  totalCount: number;
  completedCount: number;
  /** @nullable */
  startedAt: string | null;
  /** @nullable */
  updatedAt: string | null;
  /** @nullable */
  currentNoticeNumber: string | null;
  notices: BidNoticeResult[];
  searchResults: BidSearchResult[];
  keywords: string[];
  /** @nullable */
  reusedFromJobId: string | null;
  summary: BidCollectionSummary;
  /** @nullable */
  error: string | null;
}

export interface ErrorResponse {
  error: string;
}