# lib/api-zod/src/generated/types/bidAttachmentResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BidAttachmentResultDownloadStatus } from './bidAttachmentResultDownloadStatus';

export interface BidAttachmentResult {
  fileName: string;
  url: string;
  priority: boolean;
  downloadStatus: BidAttachmentResultDownloadStatus;
  /** @nullable */
  downloadError: string | null;
  attempts: number;
  parsedFileCount: number;
  parseFailureCount: number;
  keywordFound: boolean;
}