# lib/api-zod/src/generated/types/bidNoticeResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BidAttachmentResult } from './bidAttachmentResult';
import type { BidNoticeResultStatus } from './bidNoticeResultStatus';

export interface BidNoticeResult {
  noticeNumber: string;
  /** @nullable */
  noticeName: string | null;
  requestedOrder: string;
  /** @nullable */
  resolvedOrder: string | null;
  /** @nullable */
  orderMatched: boolean | null;
  lookupAttempts: number;
  status: BidNoticeResultStatus;
  attachmentCount: number;
  downloadedCount: number;
  failedCount: number;
  keywordHitCount: number;
  parseFailureCount: number;
  attachments: BidAttachmentResult[];
  /** @nullable */
  error: string | null;
}