# lib/api-zod/src/generated/types/bidSearchResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BidSearchResultAwardStatus } from './bidSearchResultAwardStatus';
import type { BidSearchResultDownloadStatus } from './bidSearchResultDownloadStatus';
import type { BidSearchResultParseStatus } from './bidSearchResultParseStatus';
import type { BidSearchResultResultStatus } from './bidSearchResultResultStatus';

export interface BidSearchResult {
  noticeNumber: string;
  /** @nullable */
  noticeName: string | null;
  attachmentFileName: string;
  downloadStatus: BidSearchResultDownloadStatus;
  /** @nullable */
  downloadError: string | null;
  parseStatus: BidSearchResultParseStatus;
  /** @nullable */
  parseError: string | null;
  resultStatus: BidSearchResultResultStatus;
  retryCount: number;
  keywordFound: boolean;
  foundKeywords: string[];
  fileName: string;
  /** @nullable */
  sheet: string | null;
  /** @nullable */
  page: number | null;
  location: string;
  surroundingText: string;
  originalText: string;
  itemSpecification?: string;
  itemName?: string;
  itemQuantity?: string;
  itemUnit?: string;
  itemAmount?: string;
  bidderName?: string;
  awardIdentifier?: string;
  awardDate?: string;
  awardAmount?: string;
  bidderAddress?: string;
  bidderPhone?: string;
  budgetAmount?: string;
  estimatedAmount?: string;
  baseAmount?: string;
  constructionOverview?: string;
  awardStatus?: BidSearchResultAwardStatus;
}